+++ CHANGELOG.rst
@@ -17,6 +17,24 @@ Possible types of changes are:
 - ``Security`` in case of vulnerabilities
 
 
+Unreleased
+----------
+
+Deprecated
+''''''''''
+- support for unresolved output variables in ``session.evaluate``, use ``sessions.solve`` for that purpose
+- support for arguments of type Variable in ``Op.__call__``, use ``Op.op`` instead
+
+Changed
+'''''''
+- simplify ``op`` decorator
+- simplify default implementation of ``Op.__repr__``
+
+Added
+'''''
+- method ``Op.op`` to be used in place of direct invocation of an Op instance when building a graph
+
+
 1.0.1 - 07.10.2020
 ------------------